simplifying code
