+++ src/Digdir.Domain.Dialogporten.WebApi/appsettings.Development.json
@@ -77,7 +77,6 @@
   },
   "CostManagement": {
     "Enabled": true,
-    "QueueCapacity": 1000,
-    "EnableQueueMonitoring": true
+    "QueueCapacity": 1000
   }
 }
