+++ src/Digdir.Domain.Dialogporten.WebApi/appsettings.prod.json
@@ -62,7 +62,6 @@
   },
   "CostManagement": {
     "Enabled": true,
-    "QueueCapacity": 500000,
-    "EnableQueueMonitoring": true
+    "QueueCapacity": 500000
   }
 }
