+++ src/Digdir.Domain.Dialogporten.WebApi/appsettings.staging.json
@@ -65,7 +65,6 @@
   },
   "CostManagement": {
     "Enabled": true,
-    "QueueCapacity": 100000,
-    "EnableQueueMonitoring": true
+    "QueueCapacity": 100000
   }
 }
