+++ src/Digdir.Domain.Dialogporten.WebApi/appsettings.test.json
@@ -62,7 +62,6 @@
   },
   "CostManagement": {
     "Enabled": true,
-    "QueueCapacity": 10000,
-    "EnableQueueMonitoring": true
+    "QueueCapacity": 10000
   }
 }
